Add files via upload
his repository contains code and data for a project on employee attrition prediction. The project uses a variety of data visualization techniques and machine learning models to predict which employees are most likely to leave their jobs. The repository includes the following files:

Employment Attrition .ipynb: A Jupyter Notebook that contains the code for the project.
data/Employee-attrition.csv: The dataset used for the project.

The project is designed to be easy to follow and reproduce. The Jupyter Notebook provides detailed explanations of the code and the results. The dataset is a fictional data set, but it is representative of the types of data that can be used for employee attrition prediction. The machine learning model is a simple random forest model, but it can be easily replaced with other models.

The project is intended to be a starting point for anyone who wants to learn about employee attrition prediction. The code and data can be used to reproduce the results of the project, or they can be used to develop new models and techniques.